requirements.txt
openai>=1.30.0
pymupdf>=1.23.0
numpy>=1.24.0
streamlit>=1.35.0
python-dotenv>=1.0.0